Ignore size value from GetBlob
The internal API in containers/image returns a size...which
is an `int64` in the source probably because OCI Descriptor
type have it be an int64 too for bad historical reasons;
xref https://github.com/opencontainers/image-spec/pull/153

Anyways, the `GetBlob` wrapper API returns `-1` when
the remote registry doesn't emit `Content-Length`:
https://github.com/containers/container-libs/blob/3af9abd27f3b875122b9fedd88953e8c840d6958/image/docker/docker_client.go#L1068

And this apparently happens with at least the Quay
mirror registry.

We also noticed this problem when adding `GetRawBlob`
in https://github.com/containers/skopeo/pull/2601
which simply doesn't return a size to the client at all.

Signed-off-by: Colin Walters <walters@verbum.org>