get_raw_blob: Properly handle -1 return from GetRawBlob
Just like GetBlob, the size might not be known. In the case
of this API, one use case is being driven from e.g.
https://github.com/cgwalters/cstor-dist
which is exposing a HTTP API, so it's really useful for
us to have the size here so that server can in turn provide
it via its own `Content-Length` if available.

Signed-off-by: Colin Walters <walters@verbum.org>